move database utils inside of tifbackendutils
