+++ .github/workflows/build_geocoder.yml
@@ -14,21 +14,6 @@ jobs:
     outputs:
       zip-path: ${{ steps.set-output.outputs.zip-path }}
 
-    env:
-      DATABASE_HOST: ${{ secrets.DATABASE_HOST }}
-      DATABASE_USERNAME: ${{ secrets.DEV_DATABASE_USERNAME }}
-      DATABASE_PASSWORD: ${{ secrets.DEV_DATABASE_PASSWORD }}
-      AWS_REGION: ${{ secrets.AWS_REGION }}
-      AWS_ACCOUNT_ID: ${{ secrets.AWS_ACCOUNT_ID }}
-      AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
-      AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
-      ABLY_KEY: ${{ secrets.ABLY_KEY }}
-      AWS_LAMBDA_FUNCTION_NAME: 'lambdaSQLRoute'
-      COGNITO_CLIENT_APP_ID: ${{ secrets.COGNITO_CLIENT_APP_ID }}
-      COGNITO_USER_POOL_ID: ${{ secrets.COGNITO_USER_POOL_ID }}
-      API_ENDPOINT: 'https://623qsegfb9.execute-api.us-west-2.amazonaws.com/stagingTest'
-      SLACK_APP_ID: ${{ secrets.SLACK_APP_ID }}
-
     steps:
       - name: Checkout code
         uses: actions/checkout@v2
