+++ .github/workflows/test_geocoder.yml
@@ -6,6 +6,10 @@ on:
     branches:
       - development
 
+concurrency: 
+  group: ${{ github.workflow }}
+  cancel-in-progress: false
+  
 jobs:
   test:
     runs-on: ubuntu-latest
