+++ APILambda/jest.config.ts
@@ -9,15 +9,13 @@ export default {
   transformIgnorePatterns: [
     "/node_modules/(?!(@planetscale|node-fetch|data-uri-to-buffer|fetch-blob|formdata-polyfill))"
   ],
-  testPathIgnorePatterns: [
-    "/dist/"
-  ],
+  testPathIgnorePatterns: ["/dist/"],
   moduleNameMapper: {
     "^(\\.{1,2}/.*)\\.js$": "$1"
   },
   globalSetup: "./test/jestSetup.ts",
   // globalTeardown: "./test/jestTeardown.ts",
-  testTimeout: 60000,
+  testTimeout: 15000,
   setupFilesAfterEnv: ["./test/setupHooks.ts"],
   watchPlugins: [
     "jest-watch-typeahead/filename",
