+++ TiFBackendUtils/MySQLDriver/test/tableDefinitions.ts
@@ -1,5 +1,3 @@
-
-
 export const tableDefintionsByFamily = [
   [
     `
