+++ TiFBackendUtils/tsconfig.json
@@ -14,6 +14,7 @@
   "exclude": [
     "**/*.test.ts",
     "**/*.spec.ts",
-    "node_modules"
+    "node_modules",
+    "test"
   ]
 }
