+++ scripts/resetDB.ts
@@ -1,4 +1,5 @@
-import { resetDB } from "../APILambda/test/database.ts";
+
+import { resetDB } from "../TiFBackendUtils/MySQLDriver/test/utils.ts";
 import { conn } from "../TiFBackendUtils/index.ts";
 
 if (process.argv.includes('--run')) {
